requirements.txt
# Core dependencies for Email Scraper & Validator
scrapling[all]>=0.2.0
email-validator>=2.0.0
py3-validate-email>=1.0.5
validate_email>=1.3
pandas>=2.0.0
tqdm>=4.65.0
beautifulsoup4>=4.12.0
phonenumbers>=8.13.0
chardet>=5.0.0
psutil>=5.9.0

# Database export dependencies (optional - only needed with --export-db flag)
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Additional dependencies included with scrapling[all]:
# - playwright
# - beautifulsoup4
# - lxml
# - camoufox (stealth browser)
gspread>=5.12.0
google-auth>=2.30.0
